+++ CMakeLists.txt
@@ -1,9 +1,15 @@
-cmake_minimum_required(VERSION 3.13.4)
+if(LINUX)
+  cmake_minimum_required(VERSION 3.13.4)
+else()
+  cmake_minimum_required(VERSION 3.27.7)
+endif()
 
 set(TARGET bbb-playground)
 
-set(CMAKE_C_COMPILER "/usr/bin/clang-13")
-set(CMAKE_CXX_COMPILER "/usr/bin/clang++-13")
+if(LINUX)
+  set(CMAKE_C_COMPILER "/usr/bin/clang-13")
+  set(CMAKE_CXX_COMPILER "/usr/bin/clang++-13")
+endif()
 
 project(${TARGET})
 
@@ -11,7 +17,9 @@ add_definitions("-std=c++2a")
 
 add_executable(${TARGET} src/app/main.cpp)
 
-target_link_libraries(${TARGET} PRIVATE librobotcontrol.so.1.0.5)
+if(LINUX)
+  target_link_libraries(${TARGET} PRIVATE librobotcontrol.so.1.0.5)
+endif()
 
 target_include_directories(${TARGET} PRIVATE src/fmt)
