+++ README.md
@@ -5,3 +5,6 @@ Project to learn to use sensors in a BeagleBone Blue board.
 ## Resources
 
 [Robot Control Library](http://www.strawsondesign.com/docs/librobotcontrol/index.html)
+[What Exactly Is GPS NMEA Data?](https://www.gpsworld.com/what-exactly-is-gps-nmea-data)
+[NMEA Output Description](http://navspark.mybigcommerce.com/content/NMEA_Format_v0.1.pdf)
+[Linux Serial Ports Using C/C++](https://blog.mbedded.ninja/programming/operating-systems/linux/linux-serial-ports-using-c-cpp)
