+++ src/include/print.h
@@ -1,20 +1,11 @@
-// print.h
-
-#ifndef PRINT_H
-#define PRINT_H
-
-#include <iostream>
+#pragma once
 
 #define FMT_HEADER_ONLY
-#include <fmt/format.h>
-
+#include "format.h"
 #include "color.h"
 
 class Print
 {
-private:
-  Color palette;
-
 public:
   Print();
   ~Print();
@@ -35,27 +26,24 @@ Print::~Print()
 
 void Print::print_title(std::string label)
 {
-  std::string formattedLabel = palette.set_color(label, "cyan", "underline");
-  std::cout << formattedLabel << "\n\n";
+  fmt::print(fmt::emphasis::underline | fg(fmt::color::cyan), "{}\n\n", label);
 }
 
 void Print::print_subtitle(std::string label)
 {
-  std::string formattedLabel = palette.set_color(label, "yellow", "underline");
-  std::cout << fmt::format("\n{}\n\n", formattedLabel);
+  fmt::print(fmt::emphasis::underline | fg(fmt::color::yellow), "\n{}\n\n", label);
 }
 
 void Print::print_info(std::string label, auto value)
 {
-  std::string formattedLabel = palette.set_color(fmt::format("{}:", label), "green");
-  std::cout << fmt::format("{} {}\n", formattedLabel, value);
+  fmt::print("{} {}\n", fmt::format(fg(fmt::color::yellow_green), "{}:", label), value);
 }
 
 void Print::print_number(std::string label, auto value, std::string units)
 {
-  std::string formattedLabel = palette.set_color(fmt::format("{}:", label), "green");
-  std::string formattedUnits = palette.set_color(units, "magenta");
-  std::cout << fmt::format("{} {:.5f} {}\n", formattedLabel, value, formattedUnits);
+  fmt::print(
+      "{} {:.7f} {}\n",
+      fmt::format(fg(fmt::color::yellow_green), "{}:", label),
+      value,
+      fmt::format(fg(fmt::color::blue_violet), "{}", units));
 }
-
-#endif // PRINT_H
